# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is swagger api for Tutree.",
        "title": "Tutree Swagger API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
        "/ambassador/authenticate": {
            "post": {
                "description": "The AmbassadorRegistration function handles the process of signing up\nstudents on a website. It expects the ambassador to submit their phone number through a form.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhoneVerification"
                ],
                "summary": "This controller will handles the registration process for ambassador.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone",
                        "name": "phone-number",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/otp/send": {
            "post": {
                "description": "This controller will resend the OTP.\nThis api is taking phone number as postform and user ip from header.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhoneVerification"
                ],
                "summary": "This controller will resend the given code same as OTP.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone-number",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "description": "This controller will verify the given code same as OTP. It will also check if OTP is expired.\nThis api is taking code and studentId as postform",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhoneVerification"
                ],
                "summary": "This controller will verify the given code same as OTP. It will also check if OTP is expired",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user-id",
                        "name": "user-id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User Agent",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session": {
            "delete": {
                "description": "Function provide a feature to delete a student's session by validating the authorization token,",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "This function handles to delete a session of a students by session id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Session ID",
                        "name": "session-id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/book-demo": {
            "post": {
                "description": "This controller will book the demo session for student.\nThis api is taking student name, parent's email, grade and session date time as postform.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DemoSessionBooking"
                ],
                "summary": "This controller funnction handles the booking of a demo session for a student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session Datetime",
                        "name": "session-datetime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Student Name",
                        "name": "student-name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Student Grade",
                        "name": "grade",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "parent's Email`",
                        "name": "parent-email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/demo-reschedule": {
            "put": {
                "description": "function reschedules a student's demo session by validating the authorization token,\nconverting session details, finding an available tutor, updating the session, and responding with a success message",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "This function handles the rescheduling of a demo session for a students.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session Datetime",
                        "name": "session-datetime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Session ID",
                        "name": "session-id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/pending-book": {
            "get": {
                "description": "Function provide a feature to list the pending classes by validating the authorization token,",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "This function give the pending classes of a students which he want to reschedule later.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "This controller will book the pending session for student.\nThis api is taking student session date time,timezone and tutor id as postform.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DemoSessionBooking"
                ],
                "summary": "This controller function handles the booking of a pending session for a student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session Datetime",
                        "name": "session-datetime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time Zone",
                        "name": "timezone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Tutor ID",
                        "name": "tutor_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/reschedule-later": {
            "post": {
                "description": "Function provide a feature to reschedules later  a student's paid session by validating the authorization token,\nsession id, and delete old record of that session.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "This function handles the rescheduling session in later of a paid session for a students.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token (Bearer token)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/student/profile": {
            "put": {
                "description": "This function handles the process of updating a student's profile information. It retrieves the student's\ncurrent profile details, processes updates to fields such as name, grade, parent's email, school ID, profile picture,\nand parent's name, and then saves these updates.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "This function handles the process of updating a student's profile information.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Grade",
                        "name": "grade",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Parent's Email",
                        "name": "parent_email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Student Name",
                        "name": "student_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Parent's Name",
                        "name": "parent_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "School ID",
                        "name": "school_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Student Picture",
                        "name": "student_picture",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/student/authenticate": {
            "post": {
                "description": "The StudentRegistration function handles the process of signing up\nstudents on a website. It expects the student to submit their phone number through a form.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhoneVerification"
                ],
                "summary": "This controller will handles the registration process for students.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone",
                        "name": "phone-number",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/student/profile": {
            "get": {
                "description": "This function retrieves the student's profile based on the provided Authorization token and\nfetches the student's demo session details. It responds with a JSON object containing the student's information\nand demo session details.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "This controller will handles to fetch profile of students.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tutor/slots": {
            "get": {
                "description": "This function retrieves the available time slots for a tutor within a specified date range so that student can\nbook demo session with available tutor.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tutor"
                ],
                "summary": "This function retrieves the available time slots for a tutor within a specified date range.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/v1/session/list": {
            "get": {
                "description": "This function retrieve a list of tutor sessions based on query parameters(If not provided then it will all list).",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tutor"
                ],
                "summary": "This Api will retrieve session list.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start-date",
                        "name": "date-from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end-date",
                        "name": "date-to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "session-type",
                        "name": "session-type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "session-status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "tutor-id",
                        "name": "tutor-id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}